新模式更改完成

+++ aust-core/src/main/java/cn/edu/aust/entity/ArticleBLOBs.java
@@ -7,7 +7,7 @@ public class ArticleBLOBs extends Article {
 
     private String content;
 
-    public ArticleBLOBs(Integer id, String title, Integer userId, String nickname, String keyword, Integer catelogId, Byte totop, Byte show, Date createdate, Date modifydate, String summary, String content) {
+    public ArticleBLOBs(Integer id, String title, Integer userId, String nickname, String keyword, Integer catelogId, Integer totop, Integer show, Date createdate, Date modifydate, String summary, String content) {
         super(id, title, userId, nickname, keyword, catelogId, totop, show, createdate, modifydate);
         this.summary = summary;
         this.content = content;
